+++ src/Order/BillingAddress.php
@@ -3,10 +3,10 @@
  * @copyright Copyright (c) Wizacha
  * @license Proprietary
  */
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Wizaplace\SDK\Order;
 
-final class ShippingAddress extends Address
+final class BillingAddress extends Address
 {
 }